automation/requirements.txt
requests
schedule
paho-mqtt
